refactor(): Move LaravelCms class to lib
